♻️ refactor: extract LlamaCppService helpers into extension files
LlamaCppService.swift was at the SwiftLint type_body_length limit.
Extract private helpers into 5 extension files to stay under the limit
and make room for upcoming reloadModel functionality (#84).

Helpers are now internal (not private) since they live in separate files;
they remain module-internal.

- LlamaCppService+ChatTemplate.swift — applyChatTemplate
- LlamaCppService+Tokenization.swift — tokenize, decodePiece
- LlamaCppService+Sampler.swift — createSampler
- LlamaCppService+Prefill.swift — prefill
- LlamaCppService+Thermal.swift — throttleIfOverheating
